conservatively reexamine request at each file reference
This is the bug fix where we remove globals that
were getting stepped on when one farm server fetched
from another site on the same farm server.

We commit this in order to test it on our only
farm site. I hope it works. It is hard to be worse.