Express: Cleanup
Switched to using the listening event that was already being emitted
by the server when ready was emitted, and gave req.headers.host it's
own variable in lib/farm.coffee.

Added a test for checking that generated args are based on given args.

Exposed the starting arguments of the server in the returned object.