don't bring up textEditor when already up
With this change, if you double-click a word in
the editor, it selects it as you would expect.

It use to be that it brought up a new editor
which would not only not have your word selected, 
it would throw away any unsaved changes. Yuck.

There is still some awkward event handling when
we double-click to enlarge a photo.