add transitions to line update
I can't believe I didn't think to add these transitions
before I shot the screencast that featured txtzyme
data updating d3-line.

http://www.youtube.com/watch?v=TeOkOebXM3w

My only excuse is that I was thinking of an oscilloscope
and they don't tween, but they should.

Although d3-line looks simple, there is a lot going
on due to svg treating a polyline as a single 
object but the circles at the verticies being
distinct objects.

Let me here thank mbostock for making transition
work so well in this case. Not only does he 
dissect the polyline to effect the transitions,
he coordinates the distinctly different transition
of the circles such that the visual effect is whole.