server does not create non-existant pages
Once we thought this was a good idea. Now we leave
it to the client to handle the 404. However, there
have been a few paths (like refreshing a long url)
that can find its way into the server-side create.

Now server side returns 404 in this case too. This
is a sinatra-only issue. Express has always 404'd 
in this case.